requirements.txt
pillow
scapy
py-datastruct